+++ server/src/visual-crossing-forecast.ts
@@ -6,16 +6,17 @@ import {
   Alert, CommonConditions, CommonConditionsKeys, CurrentConditions, CurrentConditionsKeys, DailyConditions, DailyConditionsKeys,
   DailySummaryConditions, ForecastData, ForecastDataKeys, HourlyConditions, PressureTrend
 } from './shared-types';
-import { alertCleanUp, checkForecastIntegrity, filterError, hpaToInHg } from './awcs-util';
+import { alertCleanUp, checkForecastIntegrity, checksum53, filterError, hpaToInHg } from './awcs-util';
 import { clone, isNumber, push } from '@tubular/util';
 import { floor } from '@tubular/math';
 
 interface VCAlert {
+  description: string;
+  ends: string;
   event: string;
   headline: string;
-  description: string;
+  id: string;
   onset: string;
-  ends: string;
 }
 
 interface VCCommonConditions {
@@ -303,6 +304,7 @@ function convertAlerts(vcAlerts: VCAlert[]): Alert[] {
 
     alert.title = vcAlert.headline;
     alert.description = alertCleanUp(vcAlert.description);
+    alert.id = checksum53(vcAlert.id);
     alert.time = nullIfError(floor(new Date(vcAlert.onset).getTime() / 1000));
     alert.expires = nullIfError(floor(new Date(vcAlert.ends).getTime() / 1000));