+++ src/settings-dialog.ts
@@ -1062,15 +1062,24 @@ export class SettingsDialog {
     this.reloadButton.on('click', () => setTimeout(() => location.reload()));
 
     let recentHtml = '<label>Recent locations:</label>';
+    const namesInUse = new Set<string>();
 
     this.recentLocations.forEach(loc => {
-      recentHtml += `<div class="recent-location">${htmlEscape(loc.city)}<span>✕</span></div>`;
+      let name = loc.city;
+      let index = 1;
+
+      while (namesInUse.has(name)) {
+        name = `${loc.city} (${++index})`;
+      }
+
+      namesInUse.add(name);
+      recentHtml += `<div class="recent-location">${htmlEscape(name)}<span>✕</span></div>`;
     });
 
     $('.recent-locations').html(recentHtml).find('.recent-location').each((index, elem) => {
       elem.addEventListener('click', evt => {
         if ((evt.target as HTMLElement).localName === 'span') {
-          this.recentLocations.splice(index, 1);
+          this.recentLocations[index] = null;
           (evt.target as HTMLElement).parentElement.remove();
         }
         else {
@@ -1183,6 +1192,7 @@ export class SettingsDialog {
     const newSettings = new Settings();
 
     newSettings.alertFilters = this.getAlertFilters();
+    newSettings.hiddenAlerts = this.appService.getHiddenAlerts();
 
     if (!newSettings.alertFilters)
       return;
@@ -1190,6 +1200,8 @@ export class SettingsDialog {
       this.abortForUnsavedAlarmOrBadRegex(abort => abort || this.doOK());
       return;
     }
+
+    this.recentLocations = this.recentLocations.filter(loc => loc != null);
 
     newSettings.background = this.background.val() as string;
     newSettings.celsius = (this.temperature.val() as string || '').startsWith('C');