+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 3.5.0
+
+* Added ability to acknowledge and hide weather alerts after they have been seen.
+* Filtered alerts are collapsed by default in the alerts display.
+* Fixed bug caused by selecting a previous location after deleting another previous location.
+
 ## 3.4.2
 
 * That bug fix in 3.4.1? Turns out an alarm going off on the _same day_ clocks are changed could go off at the wrong time, not merely be displayed incorrectly. That's fixed now too.